[web] Move LVM setting to storage device section
Since, using @ancorgs's words, it is more

  > consistent with the future we want to go (in which LVM is part of
  > the device selection)
