+++ web/src/components/storage/ProposalSettingsSection.test.jsx
@@ -35,221 +35,8 @@ jest.mock("@patternfly/react-core", () => {
 
 let props;
 
-const vda = {
-  sid: "59",
-  type: "disk",
-  vendor: "Micron",
-  model: "Micron 1100 SATA",
-  driver: ["ahci", "mmcblk"],
-  bus: "IDE",
-  transport: "usb",
-  dellBOSS: false,
-  sdCard: true,
-  active: true,
-  name: "/dev/vda",
-  size: 1024,
-  systems: ["Windows", "openSUSE Leap 15.2"],
-  udevIds: ["ata-Micron_1100_SATA_512GB_12563", "scsi-0ATA_Micron_1100_SATA_512GB"],
-  udevPaths: ["pci-0000:00-12", "pci-0000:00-12-ata"],
-  partitionTable: { type: "gpt", partitions: [] }
-};
-
-const md0 = {
-  sid: "62",
-  type: "md",
-  level: "raid0",
-  uuid: "12345:abcde",
-  members: ["/dev/vdb"],
-  active: true,
-  name: "/dev/md0",
-  size: 2048,
-  systems: [],
-  udevIds: [],
-  udevPaths: []
-};
-
-const md1 = {
-  sid: "63",
-  type: "md",
-  level: "raid0",
-  uuid: "12345:abcde",
-  members: ["/dev/vdc"],
-  active: true,
-  name: "/dev/md1",
-  size: 4096,
-  systems: [],
-  udevIds: [],
-  udevPaths: []
-};
-
 beforeEach(() => {
   props = {};
-});
-
-describe("LVM field", () => {
-  describe("if LVM setting is not set yet", () => {
-    beforeEach(() => {
-      props.settings = {};
-    });
-
-    it("does not render the LVM switch", () => {
-      plainRender(<ProposalSettingsSection {...props} />);
-
-      expect(screen.queryByLabelText(/Use logical volume/)).toBeNull();
-    });
-  });
-
-  describe("if LVM setting is set", () => {
-    beforeEach(() => {
-      props.settings = { lvm: false };
-    });
-
-    it("renders the LVM switch", () => {
-      plainRender(<ProposalSettingsSection {...props} />);
-
-      screen.getByRole("checkbox", { name: /Use logical volume/ });
-    });
-  });
-
-  describe("if LVM is set to true", () => {
-    beforeEach(() => {
-      props.availableDevices = [vda, md0, md1];
-      props.settings = { bootDevice: "/dev/vda", lvm: true, systemVGDevices: [] };
-      props.onChange = jest.fn();
-    });
-
-    it("renders the LVM switch as selected", () => {
-      plainRender(<ProposalSettingsSection {...props} />);
-
-      const checkbox = screen.getByRole("checkbox", { name: /Use logical volume/ });
-      expect(checkbox).toBeChecked();
-    });
-
-    it("renders a button for changing the LVM settings", () => {
-      plainRender(<ProposalSettingsSection {...props} />);
-
-      screen.getByRole("button", { name: /LVM settings/ });
-    });
-
-    it("changes the selection on click", async () => {
-      const { user } = plainRender(<ProposalSettingsSection {...props} />);
-
-      const checkbox = screen.getByRole("checkbox", { name: /Use logical volume/ });
-      await user.click(checkbox);
-
-      expect(checkbox).not.toBeChecked();
-      expect(props.onChange).toHaveBeenCalled();
-    });
-
-    describe("and user clicks on LVM settings", () => {
-      it("opens the LVM settings dialog", async () => {
-        const { user } = plainRender(<ProposalSettingsSection {...props} />);
-        const settingsButton = screen.getByRole("button", { name: /LVM settings/ });
-
-        await user.click(settingsButton);
-
-        const popup = await screen.findByRole("dialog");
-        within(popup).getByText("System Volume Group");
-      });
-
-      it("allows selecting either installation device or custom devices", async () => {
-        const { user } = plainRender(<ProposalSettingsSection {...props} />);
-        const settingsButton = screen.getByRole("button", { name: /LVM settings/ });
-
-        await user.click(settingsButton);
-
-        const popup = await screen.findByRole("dialog");
-        screen.getByText("System Volume Group");
-
-        within(popup).getByRole("button", { name: "Installation device" });
-        within(popup).getByRole("button", { name: "Custom devices" });
-      });
-
-      it("allows to set the installation device as system volume group", async () => {
-        const { user } = plainRender(<ProposalSettingsSection {...props} />);
-        const settingsButton = screen.getByRole("button", { name: /LVM settings/ });
-
-        await user.click(settingsButton);
-
-        const popup = await screen.findByRole("dialog");
-        screen.getByText("System Volume Group");
-
-        const bootDeviceButton = within(popup).getByRole("button", { name: "Installation device" });
-        const customDevicesButton = within(popup).getByRole("button", { name: "Custom devices" });
-        const acceptButton = within(popup).getByRole("button", { name: "Accept" });
-
-        await user.click(customDevicesButton);
-        await user.click(bootDeviceButton);
-        await user.click(acceptButton);
-
-        expect(props.onChange).toHaveBeenCalledWith(
-          expect.objectContaining({ systemVGDevices: [] })
-        );
-      });
-
-      it("allows customize the system volume group", async () => {
-        const { user } = plainRender(<ProposalSettingsSection {...props} />);
-        const settingsButton = screen.getByRole("button", { name: /LVM settings/ });
-
-        await user.click(settingsButton);
-
-        const popup = await screen.findByRole("dialog");
-        screen.getByText("System Volume Group");
-
-        const customDevicesButton = within(popup).getByRole("button", { name: "Custom devices" });
-        const acceptButton = within(popup).getByRole("button", { name: "Accept" });
-
-        await user.click(customDevicesButton);
-
-        const vdaOption = within(popup).getByRole("row", { name: /vda/ });
-        const md0Option = within(popup).getByRole("row", { name: /md0/ });
-        const md1Option = within(popup).getByRole("row", { name: /md1/ });
-
-        // unselect the boot devices
-        await user.click(vdaOption);
-
-        await user.click(md0Option);
-        await user.click(md1Option);
-
-        await user.click(acceptButton);
-
-        expect(props.onChange).toHaveBeenCalledWith(
-          expect.objectContaining({ systemVGDevices: ["/dev/md0", "/dev/md1"] })
-        );
-      });
-    });
-  });
-
-  describe("if LVM is set to false", () => {
-    beforeEach(() => {
-      props.settings = { lvm: false };
-      props.onChange = jest.fn();
-    });
-
-    it("renders the LVM switch as not selected", () => {
-      plainRender(<ProposalSettingsSection {...props} />);
-
-      const checkbox = screen.getByRole("checkbox", { name: /Use logical volume/ });
-      expect(checkbox).not.toBeChecked();
-    });
-
-    it("does not render a button for changing the LVM settings", () => {
-      plainRender(<ProposalSettingsSection {...props} />);
-
-      const button = screen.queryByRole("button", { name: /LVM settings/ });
-      expect(button).toBeNull();
-    });
-
-    it("changes the selection on click", async () => {
-      const { user } = plainRender(<ProposalSettingsSection {...props} />);
-
-      const checkbox = screen.getByRole("checkbox", { name: /Use logical volume/ });
-      await user.click(checkbox);
-
-      expect(checkbox).toBeChecked();
-      expect(props.onChange).toHaveBeenCalled();
-    });
-  });
 });
 
 describe("Encryption field", () => {
